requirements.txt
Django==4.1.5
django-ninja==0.20.0
djangorestframework==3.14.0